backend/requirements.txt
Django==4.2.4
django-admin-rangefilter==0.11.0
django-admin-autocomplete-filter==0.7.1
django-admin-list-filter-dropdown==1.0.3
django-debug-toolbar==4.2.0
drf-spectacular==0.26.4
djangorestframework-simplejwt==5.3.0
django-filter==23.2
factory-boy==3.3.0
requests==2.31.0
psycopg2-binary==2.9.7
pillow==10.0.0
docxtpl==0.16.7
htmldocx==0.0.6
googletrans==3.1.0a0
openai==0.27.9
sentry-sdk==1.29.2
django-import-export==3.2.0
gunicorn==21.2.0
celery==5.2.7
redis==5.0.1
django-redis==5.4.0
openpyxl==3.1.2